Introducing the new Grid type.
The Grid type is intended for use in situations where you want to
be able to layout widgets in an equitable space-sharing
arrangement, in much the same way the top-level layout works.

In fact, as a part of this change, the top-level layout is
implemented in terms of the Grid type.

The grid type also employs a Proprietary Difference Engine™ to
work out which elements of the grid have been updated, and apply
changes to only those places.  Even during very rapid development
cycles, only incremental changes are applied to the screen,
making the process that much easier.

As one final benefit, the refresh interval for the top-level
layout is now exposed as the `interval` query parameter.

Further documentation coming soon.